requirements.txt
werkzeug == 2.3.6
PyJWT~=2.8.0
setuptools~=60.2.0